Altered BB3 Code to run Victors in Test Robot
There is a problem in Teleop Init that results in the error 'No Robot Code' but only when I attempt to enable TeleOperated on the Driver Station. I moved the code to Auton Init and had the same problem with it so the problem is for sure in the Teleop Init code.